Download test cache file from alternative source
