+++ packages/beacon-state-transition/test/perf/altair/epoch/epoch.test.ts
@@ -7,7 +7,8 @@ import {
   CachedBeaconStateAltair,
   beforeProcessEpoch,
 } from "../../../../src";
-import {beforeValue, getNetworkCachedState, LazyValue} from "../../util";
+import {getNetworkCachedState} from "../../../utils/testFileCache";
+import {beforeValue, LazyValue} from "../../util";
 import {StateEpoch} from "../../types";
 import {altairState} from "../../params";
